+++ tslintReporter.js
@@ -4,23 +4,23 @@ var Model = require('./reporterModel'),
 
 function TSLintReporter(reportFile) {
     Model.call(this, reportFile);
-    global.selfJSR = this;
+    global.selfTSR = this;
 }
 
 inherits(TSLintReporter, Model);
 
 TSLintReporter.prototype.reporter = function(errors, file) {
-    var fileNbViolations = global.selfJSR.openFileIssues(file, null, /^(\s+)?\n$/gm),
+    var fileNbViolations = global.selfTSR.openFileIssues(file, null, /^(\s+)?\n$/gm),
         errorCount = errors.length,
         d = (new Date()).getTime(),
         self = this,
         severity = 'INFO';
 
     errors.forEach(function(error, index) {
-        global.selfJSR.nbViolations[global.selfJSR.INFO]++;
-        fileNbViolations[global.selfJSR.INFO]++;
+        global.selfTSR.nbViolations[global.selfTSR.INFO]++;
+        fileNbViolations[global.selfTSR.INFO]++;
 
-        fs.appendFileSync(global.selfJSR.reportFile, '{\n\t\t"line" : ' + error.startPosition.line + ',\n\t\t' +
+        fs.appendFileSync(global.selfTSR.reportFile, '{\n\t\t"line" : ' + error.startPosition.line + ',\n\t\t' +
             '"col" : ' + error.startPosition.character + ',\n\t\t' +
             '"message" : "' + error.failure.replace(/["']/g, '\'') + '",\n\t\t' +
             '"description" : "' + error.failure.replace(/["']/g, '\'') + '",\n\t\t' +
@@ -29,6 +29,6 @@ TSLintReporter.prototype.reporter = function(errors, file) {
             '"reporter" : "tslint",\n\t\t' +
             '"creationDate" : ' + d + '\n\t\t' + ((index < errorCount - 1) ? '},' : '}'));
     });
-    global.selfJSR.closeFileIssues(fileNbViolations);
+    global.selfTSR.closeFileIssues(fileNbViolations);
 };
 module.exports = TSLintReporter;