+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+<a name="2.0.4"></a>
+## 2.0.4 (2016-05-20)
+
+### Features
+
+* rulesFile option
+
+### Fixes
+
+* Typescript reporter issue
+
 <a name="2.0.3"></a>
 ## 2.0.3 (2016-05-17)